Middle of Interlevel. It's less ugly.
